[FIX] Rearranging files
