+++ pet/components/ConfirmDialog/index.tsx
@@ -8,6 +8,7 @@ import DialogTitle from '@mui/material/DialogTitle';
 import Slide from '@mui/material/Slide';
 import { TransitionProps } from '@mui/material/transitions';
 import { fira_sans_600 } from '../../core/theme/theme';
+import { CustomDialogProps } from '../CustomDialog/type';
 
 const Transition = React.forwardRef(function Transition(
     props: TransitionProps & {
@@ -18,17 +19,7 @@ const Transition = React.forwardRef(function Transition(
     return <Slide direction="up" ref={ref} {...props} />;
 });
 
-type ConfirmDialogProps = {
-    open: boolean;
-    setOpen: React.Dispatch<React.SetStateAction<boolean>>;
-    header: string;
-    description?: string;
-    cancelText?: string;
-    confirmText: string;
-    handleConfirm: React.MouseEventHandler<HTMLButtonElement>;
-};
-
-export default function ConfirmDialog(props: ConfirmDialogProps) {
+export default function ConfirmDialog(props: CustomDialogProps) {
     const { open, setOpen, header, description, cancelText, confirmText, handleConfirm } = props;
 
     const handleClose = () => {
