+++ pet/components/CustomInput/type.ts
@@ -47,38 +47,3 @@ export const CustomTextField = styled(TextField)({
     },
 });
 
-export const CustomPinkTextField = styled(TextField)({
-        
-    '& label.Mui-focused': {
-        color: '#472F05',
-    },
-    '& label': {
-        fontFamily: fira_sans_condensed.style.fontFamily
-    },
-    '& .MuiInput-underline:after': {
-        borderBottomColor: '#B2BAC2',
-    },
-    '& .MuiInputBase-input': {
-        fontFamily: fira_sans_condensed.style.fontFamily
-    },
-    '& .MuiInputBase-input:hover': {
-        backgroundColor: '#F3DDD1'
-    },
-    '& .MuiOutlinedInput-root': {
-        '& fieldset': {
-            borderRadius: 0,
-            border: '1px solid #472F05',
-            boxShadow: '3px 3px #472F05',
-        },
-        '&:hover fieldset': {
-            border: '2px solid #472F05',
-        },
-        '&.Mui-focused fieldset': {
-            border: '2px solid #472F05',
-        },
-        '&.Mui-error': {
-            color: 'red',
-            boxShadow: '3px 2px #B12000'
-        }
-    },
-});
