+++ pet/core/theme/theme.ts
@@ -1,6 +1,6 @@
 'use client'
 
-import { Tabs, styled } from '@mui/material';
+import { Tabs, styled, TextField } from '@mui/material';
 import { createTheme } from '@mui/material/styles';
 import { Fira_Sans_Condensed } from 'next/font/google';
 export const fira_sans_400 = Fira_Sans_Condensed({ weight: ['400'], subsets: ['latin'] });
@@ -82,4 +82,40 @@ export const CustomTabs = styled(Tabs)({
         color: '#6A443E',
     },
 
+});
+
+export const CustomPinkTextField = styled(TextField)({
+        
+  '& label.Mui-focused': {
+      color: '#472F05',
+  },
+  '& label': {
+      fontFamily: fira_sans_600.style.fontFamily
+  },
+  '& .MuiInput-underline:after': {
+      borderBottomColor: '#B2BAC2',
+  },
+  '& .MuiInputBase-input': {
+      fontFamily: fira_sans_600.style.fontFamily
+  },
+  '& .MuiInputBase-input:hover': {
+      backgroundColor: '#F3DDD1'
+  },
+  '& .MuiOutlinedInput-root': {
+      '& fieldset': {
+          borderRadius: 0,
+          border: '1px solid #472F05',
+          boxShadow: '3px 3px #472F05',
+      },
+      '&:hover fieldset': {
+          border: '2px solid #472F05',
+      },
+      '&.Mui-focused fieldset': {
+          border: '2px solid #472F05',
+      },
+      '&.Mui-error': {
+          color: 'red',
+          boxShadow: '3px 2px #B12000'
+      }
+  },
 });
